moved measurement server protocol into new file
